Add files via upload
Numpy is a Python library used for working with arrays. In Python we have lists that serve the purpose of arrays, but they are slow to process. Numpy aims to provide an array object that is up to 50x faster than traditional Python lists. The array object in numpy is called ndarray, it provides a lot of supporting functions that make working with ndarray very easy.